#6: Connecting multiple modules.
