--- configuration-api/README.md
@@ -1,27 +1,0 @@
-<!--suppress HtmlDeprecatedAttribute -->
-<div align="center">
-    <h1>
-        <a href="https://ninjaenterprise.github.io/NinjaConfiguration/">NinjaConfigurationService: API</a>
-    </h1>
-</div>
-
-<div align="center">
-    <a href="https://github.com/NinjaEnterprise/NinjaConfiguration/blob/master/docs/translations/README_RU.md">
-        <img alt="russian-version" src="https://raw.githubusercontent.com/NinjaEnterprise/NinjaConfiguration/master/assets/languages/russian.png"/>
-    </a>
-</div>
-
-### 📖 Description
-
-___
-
-The base module contains a common interface for working with **saving/updating/getting/deleting** configuration properties 
-and DTOs, which are used at the inputs of its methods and in requests to the configuration service.
-
-### 🎫 License
-
-___
-
-**[MIT](https://github.com/NinjaEnterprise/Ninja/blob/master/LICENSE)**
-
-_Copyright ©2021, Vladislav [[Bangerok]](https://github.com/Bangerok) Kuznetsov_
