--- configuration-client/pom.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>configuration-service</artifactId>
-    <groupId>io.github.ninjaenterprise</groupId>
-    <version>1.1.0</version>
-  </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>configuration-client</artifactId>
-  <name>Ninja Services: configuration (Feign client)</name>
-  <description>OpenFeign interface for usage into services</description>
-
-  <dependencies>
-    <!-- Spring Cloud: feign -->
-    <dependency>
-      <groupId>org.springframework.cloud</groupId>
-      <artifactId>spring-cloud-starter-openfeign</artifactId>
-    </dependency>
-
-    <!-- Configuration Service: modules -->
-    <dependency>
-      <groupId>io.github.ninjaenterprise</groupId>
-      <artifactId>configuration-api</artifactId>
-    </dependency>
-  </dependencies>
-</project>
