--- configuration-jdbc/pom.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>configuration-service</artifactId>
-    <groupId>io.github.ninjaenterprise</groupId>
-    <version>1.1.0</version>
-  </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>configuration-jdbc</artifactId>
-  <name>Ninja Services: configuration (JDBC provider)</name>
-  <description>Database for JDBC backend</description>
-
-  <properties>
-    <!-- Pluggable Modules Versions -->
-    <search-library-query-dsl.version>0.6.5</search-library-query-dsl.version>
-  </properties>
-
-  <dependencies>
-    <!-- Database connection -->
-    <dependency>
-      <groupId>org.postgresql</groupId>
-      <artifactId>postgresql</artifactId>
-      <scope>runtime</scope>
-    </dependency>
-
-    <!-- Liquibase -->
-    <dependency>
-      <groupId>org.liquibase</groupId>
-      <artifactId>liquibase-core</artifactId>
-      <scope>runtime</scope>
-    </dependency>
-
-    <!-- Hibernate annotations for validation -->
-    <dependency>
-      <groupId>org.hibernate.validator</groupId>
-      <artifactId>hibernate-validator</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <!-- Mapstruct -->
-    <dependency>
-      <groupId>org.mapstruct</groupId>
-      <artifactId>mapstruct</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <!-- Spring Boot: data -->
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-data-jpa</artifactId>
-    </dependency>
-
-    <!-- Configuration Service: modules -->
-    <dependency>
-      <groupId>io.github.ninjaenterprise</groupId>
-      <artifactId>configuration-api</artifactId>
-      <scope>provided</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-      </plugin>
-
-      <plugin>
-        <groupId>com.mysema.maven</groupId>
-        <artifactId>apt-maven-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
-</project>
