--- configuration-jdbc/src/main/resources/configuration-jdbc.yml
@@ -1,14 +1,0 @@
-# Basic spring properties
-spring:
-  # Database connections
-  datasource:
-    url: ${CONFIG_DB_URL:jdbc:postgresql://localhost:5024/configuration_db}
-    username: ${CONFIG_DB_USERNAME:ninja}
-    password: ${CONFIG_DB_PASSWORD:ninja}
-    driver-class-name: ${CONFIG_DB_DRIVER:org.postgresql.Driver}
-  # Database migration
-  jpa:
-    hibernate:
-      ddl-auto: ${CONFIG_DB_DDL_AUTO:validate}
-  liquibase:
-    change-log: ${CONFIG_LIQUIBASE_PATH:classpath:db/liquibase/db.changelog-master.yml}
