--- configuration-rest/pom.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>configuration-service</artifactId>
-    <groupId>io.github.ninjaenterprise</groupId>
-    <version>1.1.0</version>
-  </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>configuration-rest</artifactId>
-  <name>Ninja Services: configuration (REST)</name>
-  <description>REST for working create/update/delete configuration properties from database</description>
-
-  <dependencies>
-    <!-- Spring Cloud: bus-refresh -->
-    <dependency>
-      <groupId>org.springframework.cloud</groupId>
-      <artifactId>spring-cloud-starter-bus-kafka</artifactId>
-    </dependency>
-
-    <!-- Spring Cloud: discovery client -->
-    <dependency>
-      <groupId>org.springframework.cloud</groupId>
-      <artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>
-    </dependency>
-
-    <!-- Configuration Service: modules -->
-    <dependency>
-      <groupId>io.github.ninjaenterprise</groupId>
-      <artifactId>configuration-client</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>io.github.ninjaenterprise</groupId>
-      <artifactId>configuration-jdbc</artifactId>
-    </dependency>
-  </dependencies>
-</project>
