--- configuration-server/src/main/resources/application.yml
@@ -1,37 +1,0 @@
-# Basic server properties
-server:
-  port: ${CONFIG_PORT:8888}
-
-# Basic spring properties
-spring:
-  # Application
-  application:
-    name: ${CONFIG_APPLICATION_NAME:configuration-service}
-  profiles:
-    active: jdbc
-
-  # Cloud connections
-  cloud:
-    bus:
-      enabled: true
-  kafka:
-    bootstrap-servers: ${KAFKA_SERVER_URI:localhost:9092}
-
-# Basic discovery client properties
-eureka:
-  client:
-    healthcheck:
-      enabled: ${DISCOVERY_HEALTH_CHECK:false}
-    registryFetchIntervalSeconds: ${DISCOVERY_REGISTRY_FETCH_INTERVAL_SECONDS:5}
-    serviceUrl:
-      defaultZone: ${DISCOVERY_SERVICE_URL:http://localhost:8761}/eureka/
-
-# Basic actuator properties
-management:
-  endpoints:
-    web:
-      exposure:
-        include: '*'
-  endpoint:
-    health:
-      show-details: always
