+++ docker/env.bat
@@ -1,6 +1,6 @@
 set APPLICATION_NAME=NinjaConfiguration
 
-set PG_DBNAME=ninja-config-db
-set PG_USER=ninja
-set PG_PASSWORD=ninja
+set PG_DBNAME=configurations_db
+set PG_USER=postgres
+set PG_PASSWORD=123456
 set PG_PORT=5025
