+++ src/main/java/io/github/ninjaenterprise/configuration/dto/PropertyDto.java
