+++ src/main/java/io/github/ninjaenterprise/configuration/dto/PropertyReturnDto.java
