+++ src/main/java/io/github/ninjaenterprise/configuration/enums/AvailableTypes.java
