+++ src/main/java/io/github/ninjaenterprise/configuration/exceptions/ConfigurationErrors.java
