+++ src/main/java/io/github/ninjaenterprise/configuration/exceptions/ConfigurationException.java
