+++ src/main/java/io/github/ninjaenterprise/configuration/persistance/model/Property.java
