+++ src/main/java/io/github/ninjaenterprise/configuration/persistance/model/PropertyCk.java
