+++ src/main/java/io/github/ninjaenterprise/configuration/service/mapper/PropertiesMapper.java
