+++ src/main/java/io/github/ninjaenterprise/configuration/util/PropertiesUtils.java
